feat(mvc/automapper): Introducing Automapper
Automapper - package to map model to viewmodel in an elegant way.
Requires extension package for support with .net core dependency injection.

Mappings are added in a profile class in fluent syntax.
Same name mappings are dealt with seamlessly.
Where attribute/field names differ, mappings have to be specified explicitly.
Reverse mappings can be specified too using method of the same name.
Make use of this in get, getbyid and post of orders.