feat(mvc/associationCtrl): Association Controller
Mapping related entities using automapper.
How to Route related entity apis. Routing /api/orders/id/items/id
Mapped OrderItem to View Model.
How to use automapper convention of fetching related entity information using
EntityName as the prefix to the entity's property in the related entity.

Example, while fetching orderItem if we need Product information, we could fetch the properties of product that we are interested in by introducing product's properties in OrderItemViewModel with Product as a prefix to it.